fix: переименовывает файлы по критериям
