--- src/components/OfferCard/OfferCardProps.ts
@@ -1,7 +1,0 @@
-import {OfferPreviewProps} from '../../types/offerProps';
-
-export interface OfferCardProps {
-  card: OfferPreviewProps;
-  className?: string;
-  onMouseEnter?: (card:OfferPreviewProps) => void;
-}
