+++ src/components/layout/layout.tsx
@@ -1,6 +1,6 @@
 import React, {ReactNode} from 'react';
-import Header from '../Header';
-import Footer from '../Footer';
+import Header from '../header/header';
+import Footer from '../footer/footer';
 import { useLocation } from 'react-router-dom';
 import {appRoute} from '../../const';
