+++ src/pages/main.tsx
@@ -1,13 +1,13 @@
 import React, {ReactNode} from 'react';
-import {OfferDataProps} from '../components/OfferCard/OfferCardProps';
-import OfferFilter from '../components/OfferFilter';
-import OfferList from '../components/OfferList';
+import OfferFilter from '../components/offer-filter/offer-filter';
+import OfferList from '../components/offer-list/offer-list';
+import {OfferPreviewProps} from '../types/offer-props';
 
 interface MainProps {
-  offerData: OfferDataProps[];
+  offerList: OfferPreviewProps[];
 }
 
-function Main({offerData}:MainProps):ReactNode {
+function Main({offerList}:MainProps):ReactNode {
 
   return (
     <main className="page__main page__main--index">
@@ -52,10 +52,10 @@ function Main({offerData}:MainProps):ReactNode {
         <div className="cities__places-container container">
           <section className="cities__places places">
             <h2 className="visually-hidden">Places</h2>
-            <b className="places__found">{offerData.length} places to stay in Amsterdam</b>
+            <b className="places__found">{offerList.length} places to stay in Amsterdam</b>
             <OfferFilter />
             <div className="cities__places-list places__list tabs__content">
-              <OfferList offerData={offerData} className='cities'/>
+              <OfferList offerList={offerList} className='cities'/>
             </div>
           </section>
           <div className="cities__right-section">
