+++ src/types/offer-props.ts
